requirements.txt
fastapi
uvicorn[standard]
facenet-pytorch
torch
torchvision
Pillow==10.2.0
numpy>=1.24,<2.0
requests
tqdm